Separating Vertex and Vec4 + Adding Documentation
DDE-31
