+++ src/Graphics/Vertex/CMakeLists.txt
@@ -1,1 +1,5 @@
-set(VERTEX_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/Vertex.cpp PARENT_SCOPE)
+set(VERTEX_SOURCES 
+  ${CMAKE_CURRENT_SOURCE_DIR}/Vertex.cpp 
+  ${CMAKE_CURRENT_SOURCE_DIR}/Vec4.cpp 
+  PARENT_SCOPE)
+
